Renewing sorting algorithms
