+++ pom.xml
@@ -4,7 +4,7 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>org.example</groupId>
+    <groupId>org.sortingAlgorithms</groupId>
     <artifactId>sorting</artifactId>
     <version>1.0-SNAPSHOT</version>
 
@@ -12,10 +12,14 @@
         <maven.compiler.source>19</maven.compiler.source>
         <maven.compiler.target>19</maven.compiler.target>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+
+        <!-- Defining versions of dependencies as properties -->
+        <junit.jupiter.version>5.9.2</junit.jupiter.version>
+        <hamcrest.version>2.2</hamcrest.version>
+        <commons.lang.version>2.5</commons.lang.version>
     </properties>
 
     <dependencies>
-
         <dependency>
             <groupId>org.junit.jupiter</groupId>
             <artifactId>junit-jupiter</artifactId>
@@ -29,12 +33,12 @@
             <version>2.2</version>
             <scope>test</scope>
         </dependency>
+
         <dependency>
             <groupId>commons-lang</groupId>
             <artifactId>commons-lang</artifactId>
             <version>2.5</version>
         </dependency>
-
     </dependencies>
 
 </project>
