+++ src/test/java/org/sortingAlgorithms/BucketSortingTest.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.sortingAlgorithms;
 
 import org.junit.jupiter.api.Test;
 
@@ -9,12 +9,11 @@ import static org.hamcrest.MatcherAssert.assertThat;
 class BucketSortingTest {
     double[] sorted = {-45.5, -1.1, 5., 55., 805.};
     double[] unsorted = {55., 5., 805., -1.1, -45.5};
-    int numberOfBuckets = 4;
 
     @Test
     void sortSimpleArray(){
         BucketSorting bucketSorting = new BucketSorting();
-        assertThat(sorted, is(bucketSorting.bucketSort(unsorted, numberOfBuckets)));
+        assertThat(sorted, is(bucketSorting.execute(unsorted)));
     }
 
 }
